+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Features
+
+- Change `progress_callback` to return NextKeyword and `search` to return only Location
+- Update `search` to return Location bytes instead of IndexedValues
+
+### Testing
+
+- Add progress_callback tests
+
 ## [2.0.0] - 2023-01-13
 
 ### Changed